User 0 balance on create
