+++ Yum/src/main/java/org/bootcamp/tasks/ReportEmail.java
@@ -16,15 +16,13 @@ package org.bootcamp.tasks;
 
 import java.util.Date;
 import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import javax.annotation.PostConstruct;
-import org.bootcamp.yum.data.entity.Holiday;
+import java.util.regex.Pattern; 
 import org.bootcamp.yum.data.entity.Settings;
 import org.bootcamp.yum.data.repository.DailyMenuRepository;
 import org.bootcamp.yum.data.repository.HolidaysRepository;
-import org.bootcamp.yum.data.repository.SettingsRepository;
-import org.bootcamp.yum.service.DailyMenuService;
+import org.bootcamp.yum.data.repository.SettingsRepository; 
 import org.bootcamp.yum.service.EmailService;
+import org.bootcamp.yum.service.MenusService;
 import org.joda.time.Days;
 import org.joda.time.LocalDate;
 import org.joda.time.LocalDateTime;
@@ -59,7 +57,7 @@ public class ReportEmail implements ApplicationListener<ApplicationReadyEvent> {
     DailyMenuRepository dailyMenuRepo;
     
     @Autowired
-    DailyMenuService dailyMenuService;
+    MenusService menusService;
     
     private static final Logger log = LoggerFactory.getLogger(ScheduledTasks.class);
     
@@ -81,21 +79,29 @@ public class ReportEmail implements ApplicationListener<ApplicationReadyEvent> {
         
         int deadlinedays = settings.getDeadlineDays();
         LocalDate day = new LocalDate();
-        day = day.plusDays(deadlinedays);
         
-        LocalDate maxCheckdate = (new LocalDate()).plusDays(deadlinedays);
+        if(this.holidaysRepo.findByIdHoliday(day) != null){
+            return;
+        }
+                 
+        LocalDate initDate = (day).plusDays(deadlinedays);
+//        while (this.holidaysRepo.findByIdHoliday(initDate) != null) {
+//            initDate = initDate.plusDays(1);
+//        }
+        
+        LocalDate maxCheckdate = (initDate).plusDays(deadlinedays);
         while (this.holidaysRepo.findByIdHoliday(maxCheckdate) != null) {
             maxCheckdate = maxCheckdate.plusDays(1);
         }
 
-        int checkDays = Days.daysBetween((new LocalDate()), maxCheckdate).getDays() + deadlinedays - 1;
+        int checkDays = Days.daysBetween((new LocalDate()), maxCheckdate).getDays() + deadlinedays;
         
         for (int i = 0; i <= checkDays; i++) { //newDeadlineDays
-            LocalDate dailyMenuDate = (new LocalDate()).plusDays(i);
+            LocalDate dailyMenuDate = initDate.plusDays(i);
             // if old deadline not passed and new deadline passed and dailyMenu not null, send report email
-            if (dailyMenuService.deadlinePassed(dailyMenuDate) && dailyMenuRepo.findByDate(dailyMenuDate) != null) { //
-                //System.out.println(">>>>>>>>>>>>>sending email, date: " + dailyMenuDate); 
-                eService.sendOrderSummary(dailyMenuDate);
+            if (menusService.deadlinePassed(dailyMenuDate) ) { //&& dailyMenuRepo.findByDate(dailyMenuDate) != null
+                System.out.println(">>>>>>>>>>>>>sending email, date: " + dailyMenuDate); 
+                //eService.sendOrderSummary(dailyMenuDate);
             }
         }
