+++ Yum/src/main/resources/application-dev.properties
@@ -1,8 +1,7 @@
 server.port = 8080
 
-debug=false
-logging.level.root=WARN
-logging.level.org.springframework.web=WARN
+debug=true
+logging.level.org.springframework.web=DEBUG
 logging.level.org.hibernate=ERROR
