+++ bin/application-prod.properties
@@ -1,9 +1,13 @@
 server.port = 8080
 
-debug=true
-logging.level.org.springframework.web=DEBUG
-logging.level.org.hibernate=ERROR
+#debug=true
+#logging.level.org.springframework.web=DEBUG
+#logging.level.org.hibernate=ERROR
 
+debug=false
+logging.level.root=WARN
+logging.level.org.springframework.web=WARN
+logging.level.org.hibernate=ERROR
  
 spring.datasource.url= jdbc:mysql://localhost:3306/yum\?verifyServerCertificate=false&useSSL=false&requireSSL=false&useUnicode=yes&characterEncoding=UTF-8
